Move functions to persistence module
